lombok, Util.mapOf, ResultData, 게시물서비스 적용

+++ src/main/java/com/sbs/untact/cotroller/UsrArticleController.java
@@ -1,8 +1,9 @@
 package com.sbs.untact.cotroller;
 
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+
+import javax.servlet.http.HttpSession;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -22,10 +23,15 @@ public class UsrArticleController {
 
 	@RequestMapping("/usr/article/doModify")
 	@ResponseBody
-	public ResultData doModify(Integer id, String title, String body) {
+	// http://localhost:8024/usr/article/doModify?id=1&title=제목4&body=내용4
+	public ResultData doModify(Integer id, String title, String body, HttpSession session) {
 		// String title, String body는 레퍼런스라서 입력 값?을 넣지않아도 오류 안남, null값이 들어감
 		// int는 고유?타입이라서 값을 넣지않아도 null이 될 수 없음
-		Article article = articleService.getArticle(id);
+		int loginMemberId = Util.getAsInt(session.getAttribute("loginedMemberId"), 0);
+		if (loginMemberId == 0) {
+			return new ResultData("F-2", "로그인 후 이용해주세요.");
+		}
+
 		if (id == null) {
 			return new ResultData("F-1", "게시물 번호를 입력해주세요.");
 		}
@@ -38,8 +44,16 @@ public class UsrArticleController {
 			return new ResultData("F-1", "내용을 입력해주세요.");
 		}
 
+		Article article = articleService.getArticle(id);
+
 		if (article == null) {
 			return new ResultData("F-1", "해당 게시물은 존재하지 않습니다.");
+		}
+
+		ResultData actorCanModifyRd = articleService.getActorCanModify(article, loginMemberId);
+
+		if (actorCanModifyRd.isFail()) {
+			return actorCanModifyRd;
 		}
 
 		return articleService.doModify(id, title, body);
@@ -47,21 +61,43 @@ public class UsrArticleController {
 
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
-	// http://localhost:8024
-	public ResultData doDelete(int id) {
+	// http://localhost:8024/usr/article/doDelete
+	public ResultData doDelete(Integer id, HttpSession session) {
+		int loginMemberId = Util.getAsInt(session.getAttribute("loginedMemberId"), 0);
+
+		if (loginMemberId == 0) {
+			return new ResultData("F-2", "로그인 후 이용해주세요.");
+		}
+
+		if (id == null) {
+			return new ResultData("F-1", "게시물 번호를 입력해주세요.");
+		}
+
 		Article article = articleService.getArticle(id);
 
 		if (article == null) {
 			return new ResultData("F-1", "해당 게시물은 존재하지 않습니다.");
 		}
+
+		ResultData actorCanDeleteRd = articleService.getActorCanDelete(article, loginMemberId);
+
+		if (actorCanDeleteRd.isFail()) {
+			return actorCanDeleteRd;
+		}
+
 		return articleService.deleteArticle(id);
 	}
 
 	@RequestMapping("/usr/article/doAdd")
 	@ResponseBody
 	// http://localhost:8024/usr/article/doAdd?title=제목4&body=내용4
-	public ResultData doAdd(@RequestParam Map<String, Object> param) {
+	public ResultData doAdd(@RequestParam Map<String, Object> param, HttpSession session) {
 		// String title, String body이 null이면 내용이 없는 거!!
+		int loginMemberId = Util.getAsInt(session.getAttribute("loginedMemberId"), 0);
+		if (loginMemberId == 0) {
+			return new ResultData("F-2", "로그인 후 이용해주세요.");
+		}
+
 		if (param.get("title") == null) {
 			return new ResultData("F-1", "제목을 입력해주세요.");
 		}
@@ -70,16 +106,26 @@ public class UsrArticleController {
 			return new ResultData("F-1", "내용을 입력해주세요.");
 		}
 
+		param.put("MemberId", loginMemberId);
+
 		return articleService.doAdd(param);
 	}
 
 	@RequestMapping("/usr/article/detail")
 	@ResponseBody
-	// http://localhost:8024
-	public Article showDetail(int id) {
+	// http://localhost:8024/usr/article/detail
+	public ResultData showDetail(Integer id) {	
+		if (id  == null) {
+			return new ResultData("F-1", "제목을 입력해주세요.");
+		}
 
-		Article article = articleService.getArticle(id);
-		return article;
+		Article article = articleService.getForPrintArticle(id);
+		
+		if (article == null) {
+			return new ResultData("F-1", "해당 게시물은 존재하지 않습니다.");
+		}
+
+		return new ResultData("S-1", "게시물을 상세보기 입니다.", "article", article);
 	}
 
 	@RequestMapping("/usr/article/list")
@@ -101,12 +147,12 @@ public class UsrArticleController {
 		if (searchKeyword != null) {
 			searchKeyword = searchKeyword.trim();
 		} // 불필요한 뛰어쓰기 같은거는 필터링하고 검색
-		
+
 		if (searchKeyword == null) {
 			searchKeywordType = null;
 		}
 
-		return articleService.getArticleList(searchKeywordType, searchKeyword);
+		return articleService.getForPrintArticleList(searchKeywordType, searchKeyword);
 	}
 
 }
